feat: add new functionality to read in shape files
This section of code scrapes through a folder of generated shapefiles 
from mapping on 3d outcrops. It then takes all the shapefile data, which 
should have a consistent set of attributes, and combines them into a 
large data frame, summarized by several measurements we are interested 
in making.

What still needs to be done is incorporate more measurements from the 
xyz data, then ensure that the generated data has the same columns as 
the main data table for the project, and unifying them.