feat: add way to include data about whole model and join with field data.
two related tasks. first was to make it so that when measurements made 
on the model-derived shapefiles were joined with field data, they had 
the right fields to be joined on. 

the second was to find an easy way to include information that was the 
same across all shapefile shapes. This was accomplished with a separate 
file in the model data directory, which gets read in and copied across 
ll rows.