feat: changes to EPC/EAH model and accessory functions.
The original model as built by E Chamberlin and Hajek is unstable and 
buggy, here I modify the methods to provide stability over a larger 
range, as well as keeping a copy of the original script. 

I have also changed the edge handling functions, added a plotting 
routine, and a wrapper function to run many cases, and store the 
results.