feat: make manuscript figure like agu plots by scavenging components.
The goal is to have one plot showing the data that I collected in the 
Piceance Basin. The idea is that in the manuscript we want to contrast 
the not-very-different slope and depth measurements with the very 
different bar preservation. So we present these as a grid of 9 
histograms.

I have made it so that we have the bar preservation data and depth and 
slope preservation in different datasets, but we are plotting them all 
the same in a faceted arrangement, while allowing the axes to be free.

Finally, put this figure in the manuscript and copy it over using the 
makefile.